Refactoring: Create/extract an ACL test base class
The new parent class (AclTestBase) can be extended to avoid code duplication in the future (e.g. when a test for te AclManagement class will be added)
